% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thaipdf.R
\name{thaipdf_book}
\alias{thaipdf_book}
\title{Convert R Markdown to a PDF book with Thai Language Compatibility}
\usage{
thaipdf_book(...)
}
\arguments{
\item{...}{Arguments passed to \href{https://pkgs.rstudio.com/bookdown/reference/pdf_book.html}{\code{pdf_book()}}. You may supply any arguments of \code{pdf_book()} except
for \code{includes} and \code{latex_engine}.}
}
\description{
\strong{Thai language} supported conversion of R Markdown to a PDF after resolving the special tokens of \strong{bookdown} (e.g., the tokens for references and labels) to native LaTeX commands.
It is a wrapper around \href{https://pkgs.rstudio.com/bookdown/reference/pdf_book.html}{\code{bookdown::pdf_book()}} with argument \code{base_format} set to \code{\link[=thaipdf_document]{thaipdf_document()}}.
}
\seealso{
How to use \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}, please see \href{https://bookdown.org/yihui/bookdown/latexpdf.html}{official documentation}
}
