% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{thaipdf_config_set}
\alias{thaipdf_config_set}
\alias{thaipdf_config_get}
\title{Set or Get Global thaipdf Configuration}
\usage{
thaipdf_config_set(thai_font = "TH Sarabun New")

thaipdf_config_get()
}
\arguments{
\item{thai_font}{(Character) Global Thai font to set for this package, must be a valid font in your machine.}
}
\value{
Display message to R console and return list of configurations (invisibly)
}
\description{
Set and Get global configuration for \code{{thaipdf}} package. This setting will be applied to
all templates from this package.
\itemize{
\item \code{\link[=thaipdf_config_set]{thaipdf_config_set()}} for setting \code{thaipdf} global configuration
\item \code{\link[=thaipdf_config_get]{thaipdf_config_get()}} for viewing \code{thaipdf} global configuration
}
}
\details{
For example, If you want use Thai font as "Laksaman" for all
of your project, set \code{thai_font = "Laksaman"}.
Then, future templates created by this package will have "Laksaman" as main font.
}
\examples{
if(FALSE){
# Set Global Thai font
thaipdf_config_set(thai_font = "Laksaman")
# View Settings
thaipdf_config_get()
}
}
