% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thaipdf.R
\name{thaipdf_document}
\alias{thaipdf_document}
\title{Convert to a PDF/LaTeX document with Thai Language Compatibility}
\usage{
thaipdf_document(...)
}
\arguments{
\item{...}{Arguments passed to \href{https://pkgs.rstudio.com/rmarkdown/reference/pdf_document.html}{\code{pdf_document()}}. You may supply any arguments of \code{\link[=pdf_document]{pdf_document()}} except
for \code{includes} and \code{latex_engine}.}
}
\value{
R Markdown output format passing to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
\strong{Thai language} supported conversion of R Markdown to a PDF or LaTeX document.
It is a wrapper around \href{https://pkgs.rstudio.com/rmarkdown/reference/pdf_document.html}{\code{rmarkdown::pdf_document()}}.
}
\details{
To achieve Thai language compatibility for \LaTeX,
This function injects preamble fragment of Thai \LaTeX typesetting into the preamble of output \LaTeX via \href{https://pkgs.rstudio.com/rmarkdown/reference/includes.html}{\code{includes}} argument,
and set \code{latex_engine} to "xelatex".
}
\examples{
if(FALSE){
 library(rmarkdown)

 # Simple Conversion
 render("input.Rmd", output_format = thaipdf::thaipdf_document())

 # Render with Table of Contents and set font size to 10pt
 render("alone/alone.Rmd",
        output_format = thaipdf::thaipdf_document(
          toc = TRUE,
          pandoc_args = pandoc_metadata_arg("fontsize", "10pt")
        ))
}
}
\seealso{
How to use \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}, please see \href{https://bookdown.org/yihui/rmarkdown/pdf-document.html#other-features}{official documentation}.
}
