% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_thai_preamble.R
\name{ui_inform_yaml}
\alias{ui_inform_yaml}
\title{Print what TODO about R Markdown YAML to the R console}
\usage{
ui_inform_yaml(in_header = "path-to-preamble.tex")
}
\arguments{
\item{in_header}{Text to \code{in_header} fields of YAML}
}
\value{
A character
}
\description{
Print what TODO about R Markdown YAML to the R console
}
