% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_thai_preamble.R
\name{use_thai_preamble}
\alias{use_thai_preamble}
\title{Create a \LaTeX Preamble for Thai Language in R Markdown}
\usage{
use_thai_preamble(
  name = "thai-preamble.tex",
  thai_font = "TH Sarabun New",
  ignore = FALSE,
  open = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{name}{(Character) Thai \LaTeX preamble file name or path of file to create, relative to root of active project. Defaults is \code{thai-preamble.tex}}

\item{thai_font}{(Character) Name of the Thai font to use, i.e., "TH Sarabun New" (default), "Laksaman".}

\item{ignore}{(Logical) Should the newly created file be added to .Rbuildignore?}

\item{open}{(Logical) Open the newly created file for editing? Happens in RStudio, if applicable, or via \code{utils::file.edit()} otherwise.}

\item{overwrite}{(Logical) If file already exist, do you want to overwrite?}
}
\value{
(Invisible) A logical vector indicating if \LaTeX preamble file was modified.
}
\description{
This function creates \LaTeX preamble which is needed to render Thai language in R Markdown to a PDF document.
}
\details{
Here is the steps:
\enumerate{
\item Call \code{use_thai_preamble()}. You can set other file name by \code{name} argument and Thai font to use by \code{thai_font} argument.
\item Follow the instructions printed to R console.
\itemize{
\item File \code{thai-preamble.tex} should be created in the current working directory (by default).
\item (If not already) Create R Markdown file with \verb{pdf_document:} or \code{bookdown::pdf_document2} format.
\item Modify YAML header  in \verb{pdf_document:} or \code{bookdown::pdf_document2} option. Set \code{latex_engine} to \code{xelatex} and set to include path to the \LaTeX preamble file.
\item Add \LaTeX macro \verb{\\sloppy} to the beginning of the body of R Markdown (just after YAML header).
}
\item Write some Thai language in R Markdown then knit to PDF. It's DONE!
}
}
